+++ google/auth/__init__.py
@@ -17,7 +17,7 @@
 import logging
 
 from google.auth._default import default, load_credentials_from_file
-from google.auth._default_async import default_async
+
 
 
 __all__ = ["default", "load_credentials_from_file", "default_async"]
